refactor: current working dynamic list
